Cleanup types and file structure for TypeScript client
